refactor: 将EasyTier功能重构为独立的控制器模块 - 创建新的EasyTierController类来管理所有EasyTier相关功能 - 从PcView中提取EasyTier逻辑，实现关注点分离 - 通过VpnPermissionCallback接口改善解耦 - 提升代码可维护性和模块化程度
